handle non-existent efivar
PCR7 is generated from a combination of the efivars for SecureBoot, PK,
KEK, db, and dbx, though not all these variables are required to be set.

An unset variable still requires a hash of an empty structure with the
appropriate vendor GUID and UnicodeName, so handle non-existent
variables by creating the structure with an empty VariableData field.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>